Update solution file to include test project and configurations for x64 and x86 platforms; refactor HttpClient setup to rely on ServiceCollectionExtensions; change ClipStopAtLastLayers property type from int? to double?.

+++ StableDiffusionNet/Models/WebUIOptions.cs
@@ -25,7 +25,7 @@ namespace StableDiffusionNet.Models
         /// CLIP skip
         /// </summary>
         [JsonProperty("CLIP_stop_at_last_layers")]
-        public int? ClipStopAtLastLayers { get; set; }
+        public double? ClipStopAtLastLayers { get; set; }
 
         /// <summary>
         /// Включить HR upscaler по умолчанию
